+++ tests/index.test.js
@@ -237,8 +237,8 @@ describe('Quartzcron', () => {
             expect(c.out()).toBe('0 0 0 15/3 * ? *')
         })
         it('every weekend', () => {
-            c.everyWeekEnd(3, 15)
-            expect(c.out()).toBe('0 0 0 ? * 7-1 *')
+            c.everyWeekEnd()
+            expect(c.out()).toBe('0 0 0 ? * 7,1 *')
         })
         it('everyWeekDay => weekdays mon-fri', () => {
             c.everyWeekDay()
@@ -262,6 +262,10 @@ describe('Quartzcron', () => {
         it('betweenWeekDays', () => {
             c.betweenWeekDays(2, 5)
             expect(c.out()).toBe('0 0 0 ? * 2-5 *')
+        })
+        it('betweenWeekDays (wrong, unchanged)', () => {
+            c.betweenWeekDays(5, 2)
+            expect(c.out()).toBe('0 0 0 * * ? *')
         })
         it('betweenWeekDays with cadence', () => {
             c.betweenWeekDays(2, 5, 2)
@@ -297,6 +301,11 @@ describe('Quartzcron', () => {
             c.onLastMonthDay()
             expect(c.out()).toBe('0 0 0 L * ? *')
         })
+        it('onFirstMonthWeekDay', () => {
+            c.onFirstMonthWeekDay()
+            expect(c.out()).toBe('0 0 0 1W * ? *')
+        })
+
         it('onLastMonthWeekDay', () => {
             c.onLastMonthWeekDay()
             expect(c.out()).toBe('0 0 0 LW * ? *')
@@ -313,6 +322,7 @@ describe('Quartzcron', () => {
             c.onClosestWorkingDayToTheNMonthDay(2)
             expect(c.out()).toBe('0 0 0 2W * ? *')
         })
+        
         it('onNWeekDayOfTheMonth', () => {
             c.onNWeekDayOfTheMonth(2, 4)
             expect(c.out()).toBe('0 0 0 ? * 4#2 *')
@@ -452,16 +462,6 @@ describe('Quartzcron', () => {
                 .atHour('12 - 2 3 / 3')
                 .atMonth('3 - 1 0 / 2')
             expect(c.out()).toBe('3-30/2 15-45/5 12-23/3 * 3-10/2 ? *')
-        })
-    })
-
-    describe('describe as expected', () => {
-        let c
-        beforeEach(() => {
-            c = new Quartzcron()
-        })
-        it('default', () => {
-            expect(c.describe()).toBe('every second of every day of every year')
         })
     })